requirements.txt
Flask==2.1.2
joblib==1.2.0
keras==2.9.0
numpy==1.22.4
Pillow==9.2.0
Pillow==10.0.0
tensorflow==2.9.1
